Fix int to size_t for CSV file

+++ scan/dplscan_cuda.cpp
@@ -23,7 +23,7 @@ DPLScanCuda::DPLScanCuda() : Dwarf("DPLScanCuda") {}
 
 void DPLScanCuda::run_scan(const size_t buf_size, Meter &meter) {
   auto opts = meter.opts();
-  const int buffer_size = buf_size;
+  const size_t buffer_size = buf_size;
   const std::vector<int> host_src = helpers::make_random<int>(buffer_size);
 
   std::vector<int> expected =
